Code cleanup and added hepful commands
A first cleanup was done to the code.
In design_components, most classes are children of class
BaseDesignElement or class RectangleDesignElement. This way, all of them
share common methods for their dimentions and coordinates.

In parsers, class PracticalFormatParser is replaced with function
praactical_format_parcer and in placement, class RandomPlacer is
replaced with function random_placer. In both cases, they are expected
to be called once, therefore the usage unnecessary object was replaced
with functions.

New tcl commands were added: "list_components", "list_io_ports" and
"list_rows", all three of them are printing the coresponding data from
the design.