First implementation of maze routing and various changes
Fist implementation of Lee's maze routing with function and tcl command
"maze_routing". In this first implementation, the bins are filled based
on the space the nets occupy.

Bins class was added to the Design class and in the design_components
module. In various parts, changes were made for the usage of class list
instead of List.

Todo:
- Implement changes to maze routing for storing the final shape of the
net
- Add logging to the tool
- Add the ability to run scripts