Command line arguments and logging
Added command line arguments to the tool, for input script and
debugging. Arguments "-i" or "--input" are used to insert tcl script.
Changes were made to the tcl interpreter for it to evaluate the tcl
commands, based on the commands used for the tool.

Added logging to the project. Almost all prints are shown with
designation as info or error.

Changes on tcl commands were made. Tcl commands load_components" &
"save_components" were renamed to "load_components_coords" and
"save_components_coords". Tcl command "design_info" was added, which
shows desgin name, as well as information about the core and the bins.

Todo:
- Add logging file
- GUI in different thread
- Expand on net geometry after maze routing
- Timberwolf