Initial take on routing
Routed nets are created during "maze_routing". The routed net is stored
as N-ary trees, where the the root is the source and the endpoints are
the net drains. The changes on the direction of the net are stored as
points (class Netpoint) and are the nodes between source and drain.
The current routing supports cases where a point to connect exists in
the net topology.

TODO:
- Expand the net topology for points that do not exist, where a
connections is to be made to the net on a bin where the net goes but
there is no point to be connected there
- Code cleaning