Routing based on Lee's Algorithm & complementary changes
During "maze_routing", the routed nets, stored as N-ary trees are
created. The additions on this commit are that new points, between
endpoints or other points are created in order to connect the root
(source) with the various endpoints (drains). Further expansion to the
algorithm is needed for when points are created from the backtrace, but
it is not necessary for the current iteration of the Algorithm.

In class "Design", the boolean field "isRouted" was added as an
indicator for routing to not be rerun unnecessarily and for the changes
on the GUI.

Changes were made to the GUI. New buttons were added in order to show
the nets or the highlighted net either with its P2P connections or the
final routed net topology. Additionally, a heatmap for the bins was
added, connencted to its dedicated button.

New tcl commands were added. Command "net_info" shows information about
the given net & if the routing was finished, it displays the topology of
the net. Command "calculate_net_WL" calculates the wirelength of the
net, either doing the half-perimeter wirelength, or if the routing was
done, its wirelength based on its topology. Command "calculate_WL" uses
the functionality of "calculate_net_WL", but for all the nets of the
design.

TODO:
- Clean GUI code
- Add Timberwolf algorithm for better placement and results
- Multithreading:
  - GUI on a different thread
  - Routing to be done in parallel !?
- Continue adding functionality, such as blockages or limits on bins