Counterclockwise routing changes to the UI
Tcl command "maze_routing" has now the optional flag '-counterclockwise'
which is used to indicate the way routing will be done. In the default
execution, the options for wave propagation, backtracking and routing
are done clockwise, meaning NESW (North-East-South-West). With the new
flag, the direction is NWSE (North-West-South-East). Changes on the
direction affect the position of the new point created for connection.

A new set of bins was added to the design, called 'elementBins' and they
are used to store the positions occupied by the Components and IO Ports
of the design. They are also an indicator for the future to add
blockages and limits to the routing bins.

Modules "tcl_interpreter" & "gui" are grooped inside the new package
"ui". In the GUI, an additional button was added for showing the heatmap
of the element bins. Additionally, a first cleapup to the code was done.

TODO:
- Add timberwolf algorithm
- Blockages
- Expand on routing for the undeveloped cases
- Update the practical format parser for the newer designs
- Add option of storing the design to a file