Fix bugs on GUI & routing and creation of log file
On maze_routing there was an issue of point name numbering not
increasing and of some cases, connections were made based on the top
left corner of a component and not its center.

On GUI side, there was an issue discovered with the new practical format
files, where the scaling of the canvas was calculated with the window's
dimensions and not those of the canvas.

Added on logging a FileHandler writing on a file called 'router.log'.
This log keeps the same messages shown in the stdout