Added blockages and completed maze routing
Blockages were added to the design. They are created using tcl command
"add_blockage" & removed with command "remove_bloackage". The blockages
are stored in an array with the same size as the bins in the Design.
The mechanism of the blockages is that when a blockage is added on a
bin, routing cannot be done on that bin. This means that nets do not
pass on that bin, endpoints on that bin are not routes and nets starting
on that been are not been routed.
A new buttton was added to the GUI to show the blockages as a heatmap.

Tcl command "maze_routing_net" was created for testing and routing a
single net & command "maze_routing" is repetition of the single net
command for all the nets of the design.

Changes were made with the maze routing algorithm to facilitate the
addition of the blockages. Routing with multiple points created from
backtrace, as well as the posibility of a point existing to the net to
connect the new branch are the new posibilities that were addressed with
the expantion of the algorithm.

Todo:
- Code cleanup
- Run tests for all benchmarks