Added some tests and a new button
Added tests for some of the benchmarks. These tests need to be expanded
further for each test, depending on the muber of components, rows and
possible blockages to be addded on each one.

A new button was added to the GUI. Button "Hide Components Names" is
used to hide the name of the components. It is primarily useful for
tests with many components.

Todo:
- Change the new button to work as a toggle for the name and not just
hiding the name
- Add tests for all the benchmarks
- Expand algorithm to work not from the source, but from the center of
the bounding box