Added tests and changes for running them
Added the rest of the tests for all the benchmarks. It is needed for
these tests to be adapted specifically for each benchmark.

In GUI, button "Hide Comp. Names" is renamed to "Toggle Comp. Names".
Its functionality is that of a toggle for either showing or not the
names of the components. Additionally, in the heatmap buttons, when
there are no bins created, a popup message is shown.

Tcl command "place_random" has been updated and now features the ability
for the user to add as an imput the maximum number of placement tries of
the algorithm. Default number is on 100,000.