New option for maze routing
Added new option for running the maze routing algorithm. In the new
feature, the maze routing starts not form the source, but from the
center of the bounding box of the net, while the original source is
treated like any other endpoint/drain.

In TCL commands "maze_routing" and "maze_routing_net", the flag
"startRoutingFromCenter" was added in order for the new feature to be
used. The new feature is not used while using the counter clockwise
option (TODO).

In GUI, a fix was done from the previous commit where the core area was
drawnd smaller.